moved timeit_fancy to own file
